Unified integration tests using only one branch
