+++ src/test-integration/java/io/github/jopenlibs/vault/api/AuthBackendAppIdTests.java
@@ -1,9 +1,8 @@
-package io.github.jopenlibs.vault.v1_11_4.api;
+package io.github.jopenlibs.vault.api;
 
 import io.github.jopenlibs.vault.Vault;
 import io.github.jopenlibs.vault.VaultException;
-import io.github.jopenlibs.vault.response.AuthResponse;
-import io.github.jopenlibs.vault.v1_11_4.util.VaultContainer;
+import io.github.jopenlibs.vault.util.VaultContainer;
 import java.io.IOException;
 import org.junit.BeforeClass;
 import org.junit.ClassRule;
@@ -13,9 +12,9 @@ import static org.junit.Assert.assertNotNull;
 import static org.junit.Assert.assertNotSame;
 
 /**
- * Integration tests for the Username/Password auth backend.
+ * Integration tests for the AppId auth backend.
  */
-public class AuthBackendUserPassTests {
+public class AuthBackendAppIdTests {
 
     @ClassRule
     public static final VaultContainer container = new VaultContainer();
@@ -23,18 +22,20 @@ public class AuthBackendUserPassTests {
     @BeforeClass
     public static void setupClass() throws IOException, InterruptedException {
         container.initAndUnsealVault();
-        container.setupBackendUserPass();
+        container.setupBackendAppId();
     }
 
     /**
-     * Test Authentication with new userpass auth backend
+     * Test Authentication with app-id auth backend
      */
     @Test
-    public void testLoginByUserPass() throws VaultException {
+    public void testLoginByAuthId() throws VaultException {
         final Vault vault = container.getVault();
-        final AuthResponse response = vault.auth()
-                .loginByUserPass(VaultContainer.USER_ID, VaultContainer.PASSWORD);
-        final String token = response.getAuthClientToken();
+        final String path = "app-id/login";
+        @SuppressWarnings("deprecation") // used for testing
+        final String token = vault.auth()
+                .loginByAppID(path, VaultContainer.APP_ID, VaultContainer.USER_ID)
+                .getAuthClientToken();
 
         assertNotNull(token);
         assertNotSame("", token.trim());
