+++ src/test-integration/java/io/github/jopenlibs/vault/api/AuthBackendAppRoleTests.java
@@ -1,10 +1,10 @@
-package io.github.jopenlibs.vault.v1_1_3.api;
+package io.github.jopenlibs.vault.api;
 
 import io.github.jopenlibs.vault.Vault;
 import io.github.jopenlibs.vault.VaultConfig;
 import io.github.jopenlibs.vault.VaultException;
 import io.github.jopenlibs.vault.response.LogicalResponse;
-import io.github.jopenlibs.vault.v1_1_3.util.VaultContainer;
+import io.github.jopenlibs.vault.util.VaultContainer;
 import java.io.IOException;
 import org.junit.BeforeClass;
 import org.junit.ClassRule;
