+++ src/test-integration/java/io/github/jopenlibs/vault/api/AuthBackendDatabaseTests.java
@@ -1,11 +1,11 @@
-package io.github.jopenlibs.vault.v1_11_4.api;
+package io.github.jopenlibs.vault.api;
 
 import io.github.jopenlibs.vault.Vault;
 import io.github.jopenlibs.vault.VaultException;
 import io.github.jopenlibs.vault.api.database.DatabaseRoleOptions;
 import io.github.jopenlibs.vault.response.DatabaseResponse;
-import io.github.jopenlibs.vault.v1_11_4.util.DbContainer;
-import io.github.jopenlibs.vault.v1_11_4.util.VaultContainer;
+import io.github.jopenlibs.vault.util.DbContainer;
+import io.github.jopenlibs.vault.util.VaultContainer;
 import java.io.IOException;
 import java.util.ArrayList;
 import java.util.List;
