+++ src/test-integration/java/io/github/jopenlibs/vault/api/AuthBackendTokenTests.java
@@ -1,4 +1,4 @@
-package io.github.jopenlibs.vault.v1_11_4.api;
+package io.github.jopenlibs.vault.api;
 
 import io.github.jopenlibs.vault.Vault;
 import io.github.jopenlibs.vault.VaultException;
@@ -6,7 +6,7 @@ import io.github.jopenlibs.vault.api.Auth.TokenRequest;
 import io.github.jopenlibs.vault.json.Json;
 import io.github.jopenlibs.vault.response.AuthResponse;
 import io.github.jopenlibs.vault.response.LookupResponse;
-import io.github.jopenlibs.vault.v1_11_4.util.VaultContainer;
+import io.github.jopenlibs.vault.util.VaultContainer;
 import java.io.IOException;
 import java.nio.charset.StandardCharsets;
 import java.util.Arrays;
