+++ src/test-integration/java/io/github/jopenlibs/vault/api/DebugTests.java
@@ -1,9 +1,9 @@
-package io.github.jopenlibs.vault.v1_11_4.api;
+package io.github.jopenlibs.vault.api;
 
 import io.github.jopenlibs.vault.Vault;
 import io.github.jopenlibs.vault.VaultException;
 import io.github.jopenlibs.vault.response.HealthResponse;
-import io.github.jopenlibs.vault.v1_11_4.util.VaultContainer;
+import io.github.jopenlibs.vault.util.VaultContainer;
 import java.io.IOException;
 import junit.framework.TestCase;
 import org.junit.Before;
