+++ src/test-integration/java/io/github/jopenlibs/vault/api/SealTests.java
@@ -1,8 +1,8 @@
-package io.github.jopenlibs.vault.v1_11_4.api;
+package io.github.jopenlibs.vault.api;
 
 import io.github.jopenlibs.vault.VaultException;
 import io.github.jopenlibs.vault.response.SealResponse;
-import io.github.jopenlibs.vault.v1_11_4.util.VaultContainer;
+import io.github.jopenlibs.vault.util.VaultContainer;
 import java.io.IOException;
 import org.junit.BeforeClass;
 import org.junit.ClassRule;
